Dead End 1 could not remove sizes from the types definitions
The plan was to remove the sizes of var size fields and
deduce the sizes from the offsets.

The problem is that we are not writing the offsets in the
order of we write the fields(in reflection we do).
But when user is writing the Serializer itself, we write
in the order that user does but we put the indexes sorted
by the name of the field.

This prevent us to deduce the lengths from offsets because
they are not monotonically increasing anymore.

A second things is code is getting complicated pretty quickly.
We may live with it, but the reason that I stopeed following is to give freedom to users
that writes the serializers. They should be able to write the
fields in serializers in any order. This has bit us before in
Portable. Downside is we are paying extra 4 bytes per var size field.