Dead End 2 Can not support reflections with Generics
Basically, if a Class that  we want to serialize contains things like
Collection,List, Map, we can not serialize them.

The reason is that we can not create the Schema by looking at the class
only. Because of type erasure the item types are removed.

When Schema does not have the exact type for a Collection it can not
read it, since it will not know which serializer to use for the items.

The only way to support this is to put the item type on the wire
 for Collections.
In this case, we will have single type for arrays on the Schema.
One can only know the type of the collection by looking at the data.

Should we pay for this to support reflection better?
Meaningless except java(assuming C# does not have type erasure) ?