Collection type on the wire
This works until some point.
Collection<Collection> can be supported but
Collection<ArrayList> will fail on write/read. We can not failfast.

Collection of arbitrary objects can not fail fast.

It feels a lot safer to drop this and continue with Collection as
sytatic sugar over arrays.
Collection of primitives may not needed.
Collection of Decimal,Time could be good.
Collection of Nested Compact could be good.

Collection of Collection/Map is not supported.